Remove MScore::pixelRatio
MuseScore's internal coordinates are expressed in units determined by mu::engraving::DPI. The only quantity that is not expressed in this unit is text size, which is expressed in points and ignores the DPI. In our score model, doubling the DPI means for example that point (10, 10) becomes (20, 20), but a text of 10pt size does not become 20pt, it remains 10pt.

When we draw on screen we apply a transform that maps our coordinates to the screen's physical coordinates, but because the text size is expressed in different units we need to apply a different scaling factor. On top of that, QFont::pointSizeF is also DPI-aware, meaning QPainter will apply its own scaling factor based on the painter's DPI to determine the resulting pixel size.

QFont::setPixelSize is not a viable solution because it only accepts integer values. Another solution would be applying an equivalent scaling transform to the painter, but that results awkward because we need to scale only the text size but not the position. In the end the better solution is using QFont::setPointSizeF. What is bad about the current implementation is that the entire program has to be aware of the painter's DPI via MScore::pixelRatio. This is not necessary because any painter implementation is aware of its own logical DPI, so the scaling factor can be applied in the Painter class and no other part of code needs to know about it.