Set internal unit to 1/100 of millimetre
In theory, this unit is entirely arbitrary, so we could also set it to 1mm (that would make debugging nicer, cause every number we see represents millimetres). However, that corresponds to a very low DPI value (25.4), which doesn't work nicely with QFontMetrics, and very small scaled point sizes when drawing, which doesn't work nicely with QPainter. This seems the better compromise.