Maybe fix a crash in Promise
Sporadically, I encountered an ASan error just when leaving the home page (by opening a score), caused by thumbnails being loaded on background threads. The stack trace seemed to indicate a double delete of Promise::Data.

This is an attempt to capture `m_data` better in order to keep it alive. Since the crash cannot be reproduced reliably, it is difficult to verify that this solves the problem. It does not seem to introduce new problems however, and the code looks nicer without this mysterious Async::call.