Fix default smufl font size
It has somehow become common knowledge that the Smufl font size corresponding to a 7 mm staff (i.e. a spatium of 1.75 mm) is 20 pt. I too thought that that was a definition. Turns out it isn't. Smufl just prescribes that the glyphs are scaled in a way that the em size of the font corresponds to a 5-line staff height, and it just so happens that 7 mm are close to 20 typographical points (7mm / 25.4 * 72 = 19.84 pt).

Instead of computing all the font metrics using the correct point size of 19.84 pt, corresponding to our 7mm default, the metrics were being computed at 20 pt, and the SPATIUM_20 constant (which means "the value of spatium corresponding to 20 pt", different from the actual spatium) was being used throughout the codebase as a correction factor to account for the difference between 19.84 and 20.0, most notably in EngravingItem::magS. In other words we were scaling down the size of all items by a correction factor instead of computing it correctly from the start.

Having removed SPATIUM_20, it was now necessary to compute the Smufl font metrics at the correct point size.