Fix line width of page margins
Makes no sense to define this width from configutation()->borderWidth() cause that's a value for UI logical pixels (and it's just equal to 1), but this painter operates in notation units determined by mu::engraving::DPI (and scales when zooming on the score). The correct thing is to define it in millimetres (0.07mm turns out to be the corresponding value).