Move score color inversion setting out of EngravingConfiguration
into NotationConfiguration.

It has evolved into a UI setting, with the possibility to follow the app theme, and as such does not fit in the engraving module anymore. The engraving module only needs to know whether a certain paint action should use inverted colors or not; that info is passed in via PaintOptions.