Remove "length" property from SLine palette items
This doesn't seem to serve any purpose, it only exists to save the length of SLines in palettes. But that length is irrelevant once the item is inserted in the score (because it's determined by layout), and it's also irrelevant in the Palette, where it should just be set to whatever value allows to draw them properly in the cell.