fix(properties): empty playback popup
Root cause: the required property `model` was not properly set in the instantiation of `NoteExpandableBlank` etc.. Apparently, in the case of Loader, properties of the Loader are not automatically assigned to required properties of the source component; they are only passed as context properties, which we can't use with `ComponentBehaviour: Bound`.
Solution: pass the property directly where the components are declared. This does require moving the component declarations into the Loader, because only there the model in question is available. That in turn required a slight refactor of `fullModel` and `getVisibleModel`.

This is combined with a refactor that eliminates PlaybackGeneralSettings, because it is the only child of PlaybackSettings, so everything can just as well be directly in PlaybackSettings.

And a tiny runtime warning fix in `NotationMultiElementView`.